delete the get-like methods in files
