+++ pyanomaly/loss/functions/__init__.py
@@ -3,4 +3,3 @@
 @contact: yuhao.cheng[at]outlook.com
 """
 from .basic_loss import *
-from .expand_loss import *
