+++ pyanomaly/networks/meta/anopcn_networks.py
@@ -18,7 +18,7 @@ from pyanomaly.networks.meta.base.commonness import (
 
 from ..model_registry import META_ARCH_REGISTRY
 
-__all__ = ['AnoPcn', 'get_model_anopcn']
+__all__ = ['AnoPcn']
 
 class ERM(nn.Module):
     def __init__(self, c_in, c_out, bilinear=False):
@@ -269,33 +269,3 @@ class AnoPcn(nn.Module):
         
         return prediction, result
 
-
-def get_model_anopcn(cfg):
-    if cfg.ARGUMENT.train.normal.use:
-        rgb_max = 1.0
-    else:
-        rgb_max = 255.0
-    if cfg.MODEL.flownet == 'flownet2':
-        from collections import namedtuple
-        from pyanomaly.networks.auxiliary.flownet2.models import FlowNet2
-        temp = namedtuple('Args', ['fp16', 'rgb_max'])
-        args = temp(False, rgb_max)
-        flow_model = FlowNet2(args)
-        flow_model.load_state_dict(torch.load(cfg.MODEL.flow_model_path)['state_dict'])
-    elif cfg.MODEL.flownet == 'liteflownet':
-        from pyanomaly.networks.auxiliary.liteflownet.models import LiteFlowNet
-        flow_model = LiteFlowNet()
-        flow_model.load_state_dict({strKey.replace('module', 'net'): weight for strKey, weight in torch.load(cfg.MODEL.flow_model_path).items()})
-    else:
-        raise Exception('Not support optical flow methods')
-
-    generator_model = AnoPcn(cfg)
-    discriminator_model = PixelDiscriminator(3, cfg.MODEL.discriminator_channels, use_norm=False)
-    # discriminator_model = NLayerDiscriminator(3)
-    
-    model_dict = OrderedDict()
-    model_dict['Generator'] = generator_model
-    model_dict['Discriminator'] = discriminator_model
-    model_dict['FlowNet'] = flow_model
-    
-    return model_dict
