+++ pyanomaly/networks/meta/ocae_networks.py
@@ -10,7 +10,8 @@ from collections import OrderedDict
 from sklearn.multiclass import OneVsRestClassifier
 from sklearn.svm import LinearSVC
 from ..model_registry import META_ARCH_REGISTRY
-__all__ = ['CAE', 'get_model_ocae']
+
+__all__ = ['CAE']
 
 @META_ARCH_REGISTRY.register()
 class CAE(nn.Module):
@@ -55,31 +56,4 @@ class CAE(nn.Module):
         output = self.decoder(latent_feature)
         return latent_feature, output, x
 
-def get_model_ocae(cfg):
-    model_dict = OrderedDict()
-    model_dict['A'] = CAE(c_in=1)
-    model_dict['B'] = CAE(c_in=1)
-    model_dict['C'] = CAE(c_in=1)
-    if cfg.MODEL.name == 'ocae':
-        import detectron2
-        from detectron2.engine import DefaultPredictor
-        from detectron2.checkpoint import DetectionCheckpointer
-        from detectron2.modeling import build_model
-        from detectron2.config import get_cfg
-        from detectron2 import model_zoo
-    else:
-        raise Exception('Not the correct the model name')
-    detector_cfg = get_cfg()
-    file_name = cfg.MODEL.detector_config
-    detector_cfg.merge_from_file(model_zoo.get_config_file(file_name))
-    detector_cfg.MODEL.ROI_HEADS.SCORE_THRESH_TEST = 0.5
-    detector_cfg.MODEL.ROI_HEADS.NMS_THRESH_TEST = 0.8
-    det_model = build_model(detector_cfg)
-    DetectionCheckpointer(det_model).load(cfg.MODEL.detector_model_path)
-    det_model.train(False)
-    model_dict['Detector'] = det_model
-
-    model_dict['OVR'] = OneVsRestClassifier(LinearSVC(random_state = 0), n_jobs=16)
-
-    return model_dict
